@@ -38,10 +38,10 @@
     "libs/**/*"
   ],
   "devDependencies": {
-    "@angular-devkit/build-angular": "19.2.14",
-    "@angular-eslint/schematics": "19.6.0",
-    "@angular/cli": "19.2.14",
-    "@angular/compiler-cli": "19.2.14",
+    "@angular-devkit/build-angular": "20.3.11",
+    "@angular-eslint/schematics": "19.8.1",
+    "@angular/cli": "20.3.11",
+    "@angular/compiler-cli": "20.3.13",
     "@babel/core": "7.24.9",
     "@babel/preset-env": "7.24.8",
     "@compodoc/compodoc": "1.1.26",
@@ -49,7 +49,7 @@
     "@electron/rebuild": "4.0.1",
     "@eslint/compat": "1.2.9",
     "@lit-labs/signals": "0.1.2",
-    "@ngtools/webpack": "19.2.14",
+    "@ngtools/webpack": "20.3.11",
     "@storybook/addon-a11y": "9.1.16",
     "@storybook/addon-designs": "9.0.0-next.3",
     "@storybook/addon-docs": "9.1.16",
@@ -82,7 +82,7 @@
     "@typescript-eslint/utils": "8.31.0",
     "@webcomponents/custom-elements": "1.6.0",
     "@yao-pkg/pkg": "6.5.1",
-    "angular-eslint": "19.6.0",
+    "angular-eslint": "20.7.0",
     "autoprefixer": "10.4.21",
     "axe-playwright": "2.2.2",
     "babel-loader": "9.2.1",
@@ -148,15 +148,15 @@
     "webpack-node-externals": "3.0.0"
   },
   "dependencies": {
-    "@angular/animations": "19.2.14",
+    "@angular/animations": "20.3.13",
     "@angular/cdk": "19.2.18",
-    "@angular/common": "19.2.14",
-    "@angular/compiler": "19.2.14",
-    "@angular/core": "19.2.14",
-    "@angular/forms": "19.2.14",
-    "@angular/platform-browser": "19.2.14",
-    "@angular/platform-browser-dynamic": "19.2.14",
-    "@angular/router": "19.2.14",
+    "@angular/common": "20.3.13",
+    "@angular/compiler": "20.3.13",
+    "@angular/core": "20.3.13",
+    "@angular/forms": "20.3.13",
+    "@angular/platform-browser": "20.3.13",
+    "@angular/platform-browser-dynamic": "20.3.13",
+    "@angular/router": "20.3.13",
     "@bitwarden/commercial-sdk-internal": "0.2.0-main.375",
     "@bitwarden/sdk-internal": "0.2.0-main.375",
     "@electron/fuses": "1.8.0",
@@ -165,7 +165,7 @@
     "@koa/router": "14.0.0",
     "@microsoft/signalr": "8.0.7",
     "@microsoft/signalr-protocol-msgpack": "8.0.7",
-    "@ng-select/ng-select": "14.9.0",
+    "@ng-select/ng-select": "20.7.0",
     "@nx/devkit": "21.6.8",
     "@nx/eslint": "21.6.8",
     "@nx/jest": "21.6.8",
